fix: fix mistakes for pull request#1
