--- src/account/account.module.ts
@@ -1,17 +1,0 @@
-import { Module } from '@nestjs/common';
-import { MongooseModule } from '@nestjs/mongoose';
-import { Account, AccountSchema } from './entities/account.entity';
-import { AccountController } from './account.controller';
-import { AccountService } from './account.service';
-
-@Module({
-    imports: [
-        MongooseModule.forFeature([
-            { name: Account.name, schema: AccountSchema },
-        ]),
-    ],
-    controllers: [AccountController],
-    providers: [AccountService],
-    exports: [AccountService],
-})
-export class AccountModule {}
