--- src/account/dto/update-amo-integration.dto.ts
@@ -1,4 +1,0 @@
-import { PartialType } from '@nestjs/mapped-types';
-import { CreateAccountDto } from './create-account.dto';
-
-export class UpdateAmoIntegrationDto extends PartialType(CreateAccountDto) {}
