--- src/custom-field/dto/update-custom.field.dto.ts
@@ -1,4 +1,0 @@
-import { PartialType } from '@nestjs/mapped-types';
-import { CreateCustomFieldDto } from './create-cutom-field.dto';
-
-export class UpdateCustomFieldDto extends PartialType(CreateCustomFieldDto) {}
