+++ src/main.ts
@@ -1,12 +1,13 @@
 import { NestFactory } from '@nestjs/core';
 import { AppModule } from './app/app.module';
 import { ConfigService } from '@nestjs/config';
+import { env } from './shared/env.enum';
 
 async function bootstrap(): Promise<void> {
     const app = await NestFactory.create(AppModule);
     const configService = app.get(ConfigService);
 
-    const port = configService.get('PORT');
+    const port = configService.get(env.Port);
 
     await app.listen(port);
 }
