+++ src/modules/account/dto/create-account.dto.ts
@@ -1,0 +1,18 @@
+import { IsBoolean, IsNumber, IsString } from 'class-validator';
+
+export class CreateAccountDTO {
+    @IsNumber()
+    public readonly accountId: number;
+
+    @IsString()
+    public readonly subdomain: string;
+
+    @IsString()
+    public readonly accessToken: string;
+
+    @IsString()
+    public readonly refreshToken: string;
+
+    @IsBoolean()
+    public readonly isInstalled: boolean;
+}
