+++ src/modules/account/dto/update-amo-integration.dto.ts
@@ -1,0 +1,6 @@
+import { PartialType, OmitType } from '@nestjs/mapped-types';
+import { CreateAccountDTO } from './create-account.dto';
+
+export class UpdateAmoIntegrationDTO extends PartialType(
+    OmitType(CreateAccountDTO, ['accountId'] as const)
+) {}
