+++ src/modules/account/enums/errors.enum.ts
@@ -1,0 +1,3 @@
+export enum envError {
+    Not_Found = 'Аккаунт с переданным accountId не был найден',
+}
