+++ src/modules/account/enums/limit.enum.ts
@@ -1,0 +1,3 @@
+export enum envLimit {
+    Limit_Account = 5,
+}
