+++ src/modules/amo-api/enums/error.enum.ts
@@ -1,0 +1,7 @@
+export enum envError {
+    Post_Access_Refresh_Tokens = 'Ошибка при запросе при получении access token и refresh token',
+    Get_Account_Data = 'Ошибка при запросе при получении данных аккаунта',
+    Get_Custom_Fields = 'Ошибка при запросе при получении кастомных полей',
+    Post_Custom_Fields = 'Ошибка при запросе при создании кастомных полей',
+    Delete_Custom_Fields = 'Ошибка при запросе при удалении кастомных полей',
+}
