+++ src/modules/custom-field/dto/update-custom.field.dto.ts
@@ -1,0 +1,4 @@
+import { PartialType } from '@nestjs/mapped-types';
+import { CreateCustomFieldDTO } from './create-cutom-field.dto';
+
+export class UpdateCustomFieldDto extends PartialType(CreateCustomFieldDTO) {}
