+++ src/modules/custom-field/types/types.ts
@@ -1,0 +1,8 @@
+export type TRequestCustomFields = {
+    accountId: number;
+    entityType: string;
+};
+
+export type TAccount_ID = {
+    accountId: number;
+};
